New super-fast send and receive implementation
I decided to split the MPI functions into a fast, unsafe variant
that only works for simple-arrays and a slightly slower variant
that can send any object. The former is intended to mimic the
MPI C bindings in speed and expressiveness. The latter should be
able to transmit any lisp object in a typesafe manner.