Change the number of processes in the test suite from 4 to 2.
The reason for this is that newer versions of OpenMPI complain about
oversubscription when the computer doesn't have as many physical cores as
requested.  One way to fix this would be to add the '--oversubscribe'
option, but I am not sure whether doing so is portable.  So the less
invasive solution is to reduce the number of processes to 2, because it is
very hard to find a modern computer that doesn't have at least two physical
cores.